collating DIS routines, removing unused
