+++ src/include/batch_request.h
@@ -405,7 +405,6 @@ extern int decode_DIS_UserCred(int, struct batch_request *);
 extern int decode_DIS_JobFile(int, struct batch_request *);
 extern int decode_DIS_CopyHookFile(int, struct batch_request *);
 extern int decode_DIS_DelHookFile(int, struct batch_request *);
-extern int decode_DIS_JobObit(int, struct batch_request *);
 extern int decode_DIS_Manage(int, struct batch_request *);
 extern int decode_DIS_DelJobList(int, struct batch_request *);
 extern int decode_DIS_MoveJob(int, struct batch_request *);
@@ -430,7 +429,6 @@ extern int decode_DIS_Cred(int, struct batch_request *);
 extern int encode_DIS_failover(int, struct batch_request *);
 extern int encode_DIS_CopyFiles(int, struct batch_request *);
 extern int encode_DIS_CopyFiles_Cred(int, struct batch_request *);
-extern int encode_DIS_JobObit(int, struct batch_request *);
 extern int encode_DIS_Register(int, struct batch_request *);
 extern int encode_DIS_TrackJob(int, struct batch_request *);
 extern int encode_DIS_reply(int, struct batch_reply *);
