--- src/lib/Libifl/dec_Authen.c
@@ -1,108 +1,0 @@
-/*
- * Copyright (C) 1994-2021 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of both the OpenPBS software ("OpenPBS")
- * and the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * OpenPBS is free software. You can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- *
- * OpenPBS is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * PBS Pro is commercially licensed software that shares a common core with
- * the OpenPBS software.  For a copy of the commercial license terms and
- * conditions, go to: (http://www.pbspro.com/agreement.html) or contact the
- * Altair Legal Department.
- *
- * Altair's dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of OpenPBS and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair's trademarks, including but not limited to "PBS™",
- * "OpenPBS®", "PBS Professional®", and "PBS Pro™" and Altair's logos is
- * subject to Altair's trademark licensing policies.
- */
-
-
-
-/**
- * @file	dec_Authen.c
- * @brief
- * 	decode_DIS_Authenticate() - decode a authentication request
- *
- *	The batch_request structure must already exist (be allocated by the
- *	caller.   It is assumed that the header fields (protocol type,
- *	protocol version, request type, and user name) have already be decoded.
- *
- *	Data items are:	unsigned int	port number
- */
-
-#include <pbs_config.h>   /* the master config generated by configure */
-
-#include <sys/types.h>
-#include "libpbs.h"
-#include "list_link.h"
-#include "server_limits.h"
-#include "attribute.h"
-#include "credential.h"
-#include "batch_request.h"
-#include "dis.h"
-
-/**
- * @brief
- *      Decode PBS batch request to authenticate based on external (non-resv-port) mechanisms.
- *      The batch request contains type and the auth data.
- *
- * @param [in] sock socket connection
- * @param [in] preq PBS bath request
- * @return in
- * @retval 0 on success
- * @retval > 0 on failure
- */
-int
-decode_DIS_Authenticate(int sock, struct batch_request *preq)
-{
-	int rc;
-	int len = 0;
-
-	memset(preq->rq_ind.rq_auth.rq_auth_method, '\0', sizeof(preq->rq_ind.rq_auth.rq_auth_method));
-	len = disrsi(sock, &rc);
-	if (rc != DIS_SUCCESS)
-		return (rc);
-	if (len <= 0) {
-		return DIS_PROTO;
-	}
-	rc = disrfst(sock, len, preq->rq_ind.rq_auth.rq_auth_method);
-	if (rc != DIS_SUCCESS)
-		return (rc);
-
-	memset(preq->rq_ind.rq_auth.rq_encrypt_method, '\0', sizeof(preq->rq_ind.rq_auth.rq_encrypt_method));
-	len = disrsi(sock, &rc);
-	if (rc != DIS_SUCCESS)
-		return (rc);
-	if (len > 0) {
-		rc = disrfst(sock, len, preq->rq_ind.rq_auth.rq_encrypt_method);
-		if (rc != DIS_SUCCESS)
-			return (rc);
-	}
-
-	preq->rq_ind.rq_auth.rq_port = disrui(sock, &rc);
-	if (rc != DIS_SUCCESS)
-		return (rc);
-
-	return (rc);
-}
