--- src/lib/Libifl/dec_ReqHdr.c
@@ -1,108 +1,0 @@
-/*
- * Copyright (C) 1994-2021 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of both the OpenPBS software ("OpenPBS")
- * and the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * OpenPBS is free software. You can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- *
- * OpenPBS is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * PBS Pro is commercially licensed software that shares a common core with
- * the OpenPBS software.  For a copy of the commercial license terms and
- * conditions, go to: (http://www.pbspro.com/agreement.html) or contact the
- * Altair Legal Department.
- *
- * Altair's dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of OpenPBS and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair's trademarks, including but not limited to "PBS™",
- * "OpenPBS®", "PBS Professional®", and "PBS Pro™" and Altair's logos is
- * subject to Altair's trademark licensing policies.
- */
-
-
-
-/**
- * @file	dec_ReqHdr.c
- * @brief
- * decode_DIS_ReqHdr() - Decode the Request Header Fields
- *	common to all requests
- *
- * @par	Fields are:
- * 			Protocol ID (unsigned integer)
- *			Protocol Version (unsigned integer)
- *			Request Type (unsignded integer)
- *			User Name (string)
- *
- * @par Returns:
- *		  -1	on EOF (end of file on first read only)
- *		   0	on success
- *		  >0	a DIS error return, see dis.h
- */
-
-#include <pbs_config.h>   /* the master config generated by configure */
-
-#include <sys/types.h>
-#include "libpbs.h"
-#include "list_link.h"
-#include "server_limits.h"
-#include "attribute.h"
-#include "credential.h"
-#include "batch_request.h"
-#include "dis.h"
-
-/**
- * @brief-
- *	Decode the Request Header Fields
- *      common to all requests
- *
- * @param[in] sock - socket descriptor
- * @param[out] preq - pointer to batch_request structure
- *
- * @return	int
- * @retval	-1    on EOF (end of file on first read only)
- * @retval	0    on success
- * @retval	>0    a DIS error return, see dis.h
- *
- */
-
-int
-decode_DIS_ReqHdr(int sock, struct batch_request *preq, int *proto_type, int *proto_ver)
-{
-	int rc;
-
-	*proto_type = disrui(sock, &rc);
-	if (rc) {
-		return rc;
-	}
-	if (*proto_type != PBS_BATCH_PROT_TYPE)
-		return  DIS_PROTO;
-	*proto_ver = disrui(sock, &rc);
-	if (rc) {
-		return rc;
-	}
-
-	preq->rq_type = disrui(sock, &rc);
-	if (rc) {
-		return rc;
-	}
-
-	return (disrfst(sock, PBS_MAXUSER+1, preq->rq_user));
-}
