--- src/lib/Libifl/enc_MsgJob.c
@@ -1,143 +1,0 @@
-/*
- * Copyright (C) 1994-2021 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of both the OpenPBS software ("OpenPBS")
- * and the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * OpenPBS is free software. You can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- *
- * OpenPBS is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * PBS Pro is commercially licensed software that shares a common core with
- * the OpenPBS software.  For a copy of the commercial license terms and
- * conditions, go to: (http://www.pbspro.com/agreement.html) or contact the
- * Altair Legal Department.
- *
- * Altair's dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of OpenPBS and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair's trademarks, including but not limited to "PBS™",
- * "OpenPBS®", "PBS Professional®", and "PBS Pro™" and Altair's logos is
- * subject to Altair's trademark licensing policies.
- */
-
-
-
-/**
- * @file	enc_MsgJob.c
- * @brief
- * encode_DIS_MessageJob() - encode a Message Job Batch Request
- *
- * @par	Data items are:
- * 			string		job id
- *			unsigned int	which file (fileopt)
- *			string		the message
- */
-
-#include <pbs_config.h>   /* the master config generated by configure */
-
-#include "libpbs.h"
-#include "pbs_error.h"
-#include "dis.h"
-
-/**
- * @brief
- *	- encode a Message Job Batch Request
- *
- * @param[in] sock - socket descriptor
- * @param[in] jobid - job id
- * @param[in] fileopt - which file
- * @param[in] msg - msg to be encoded
- *
- * @return      int
- * @retval      DIS_SUCCESS(0)  success
- * @retval      error code      error
- *
- */
-
-int
-encode_DIS_MessageJob(int sock, char *jobid, int fileopt, char *msg)
-{
-	int   rc;
-
-	if ((rc = diswst(sock, jobid) != 0) ||
-		(rc = diswui(sock, fileopt) != 0) ||
-		(rc = diswst(sock, msg) != 0))
-			return rc;
-
-	return 0;
-}
-
-/**
- * @brief
- *	-Write a python spawn request onto the wire.
- *	Each of the argv and envp arrays is sent by writing a counted
- *	string followed by a zero length string ("").  They are read
- *	by the function read_carray() in dec_MsgJob.c
- *
- * @param[in] sock - socket descriptor
- * @param[in] jobid - job id
- * @param[in] argv - pointer to argument list
- * @param[in] envp - pointer to environment variable
- *
- * @return      int
- * @retval      DIS_SUCCESS(0)  success
- * @retval      error code      error
- *
- */
-int
-encode_DIS_PySpawn(int sock, char *jobid, char **argv, char **envp)
-{
-	int	rc, i;
-	char	*cp;
-
-	if ((rc = diswst(sock, jobid)) != DIS_SUCCESS)
-		return rc;
-
-	if (argv != NULL) {
-		for (i=0; (cp = argv[i]) != NULL; i++) {
-			if ((rc = diswcs(sock, cp, strlen(cp))) != DIS_SUCCESS)
-				return rc;
-		}
-	}
-	if ((rc = diswcs(sock, "", 0)) != DIS_SUCCESS)
-		return rc;
-
-	if (envp != NULL) {
-		for (i=0; (cp = envp[i]) != NULL; i++) {
-			if ((rc = diswcs(sock, cp, strlen(cp))) != DIS_SUCCESS)
-				return rc;
-		}
-	}
-	rc = diswcs(sock, "", 0);
-
-	return rc;
-}
-
-int
-encode_DIS_RelnodesJob(int sock, char *jobid, char *node_list)
-{
-	int   rc;
-
-	if ((rc = diswst(sock, jobid) != 0) ||
-		(rc = diswst(sock, node_list) != 0))
-			return rc;
-
-	return 0;
-}
