--- src/lib/Libifl/enc_ReqExt.c
@@ -1,85 +1,0 @@
-/*
- * Copyright (C) 1994-2021 Altair Engineering, Inc.
- * For more information, contact Altair at www.altair.com.
- *
- * This file is part of both the OpenPBS software ("OpenPBS")
- * and the PBS Professional ("PBS Pro") software.
- *
- * Open Source License Information:
- *
- * OpenPBS is free software. You can redistribute it and/or modify it under
- * the terms of the GNU Affero General Public License as published by the
- * Free Software Foundation, either version 3 of the License, or (at your
- * option) any later version.
- *
- * OpenPBS is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
- * License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Commercial License Information:
- *
- * PBS Pro is commercially licensed software that shares a common core with
- * the OpenPBS software.  For a copy of the commercial license terms and
- * conditions, go to: (http://www.pbspro.com/agreement.html) or contact the
- * Altair Legal Department.
- *
- * Altair's dual-license business model allows companies, individuals, and
- * organizations to create proprietary derivative works of OpenPBS and
- * distribute them - whether embedded or bundled with other software -
- * under a commercial license agreement.
- *
- * Use of Altair's trademarks, including but not limited to "PBS™",
- * "OpenPBS®", "PBS Professional®", and "PBS Pro™" and Altair's logos is
- * subject to Altair's trademark licensing policies.
- */
-
-
-
-/**
- * @file	enc_ReqExt.c
- * @brief
- * encode_DIS_ReqExtend() - write an extension to a Batch Request
- *	The extension is in two parts:
- *		unsigned integer - 1 if an extension string follows, 0 if not
- *		character string - if 1 above
- */
-
-#include <pbs_config.h>   /* the master config generated by configure */
-
-#include "dis.h"
-
-/**
- * @brief
- *	-write an extension to a Batch Request
- *
- * @par	The extension is in two parts:
- *		unsigned integer - 1 if an extension string follows, 0 if not\n
- *		character string - if 1 above
- *
- * @param[in] sock - socket descriptor
- * @param[in] extend - string which used as extension for req
- *
- * @return      int
- * @retval      DIS_SUCCESS(0)  success
- * @retval      error code      error
- *
- */
-
-int
-encode_DIS_ReqExtend(int sock, char *extend)
-{
-	int rc;
-
-	if ((extend == NULL) || (*extend == '\0')) {
-		rc = diswui(sock, 0);
-	} else {
-		if ((rc = diswui(sock, 1)) == 0) {
-			rc = diswst(sock, extend);
-		}
-	}
-	return rc;
-}
